+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- It is possible to run an image generated with the ``cartridge pack docker``
+  command in an unprivileged Kubernetes container. It became possible, because
+  tarantool user now always has ``UID = 1200`` and ``GID = 1200``.
+
 ## [2.7.2] - 2021-03-24
 
 ### Changed
@@ -15,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated `metrics` to `0.7.1` in application template
 - Updated `cartridge-cli-extensions` to `1.1.1` in application template
 
-## Added
+### Added
 
 - Variables ``TARANTOOL_WORKDIR``, ``TARANTOOL_PID_FILE`` and
   ``TARANTOOL_CONSOLE_SOCK`` can be customized when packing in docker via
@@ -28,7 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Updated `cartridge` to `2.5.0` in application template
 
-## Fixed
+### Fixed
 
 - Added interruption of an incomplete expression when pressing
   ``Ctrl-C`` in ``cartridge enter`` command.